# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriChain Digital Agriculture Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certifications/enrolments:
    get:
      operationId: listEnrolments
      tags: [certifications]
      summary: List certification enrolments
      parameters:
        - in: query
          name: farmerId
          schema:
            type: string
        - in: query
          name: streamId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, expired, pending]
      responses:
        "200":
          description: Enrolments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CertificationEnrolment"

    post:
      operationId: createEnrolment
      tags: [certifications]
      summary: Enrol farmer in certification stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrolmentBody"
      responses:
        "201":
          description: Enrolment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificationEnrolment"

  # ============================================================
  # SURVEYS
  # ============================================================
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List commodity batches
      parameters:
        - in: query
          name: agentId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [open, locked, delivered]
      responses:
        "200":
          description: Batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"

    post:
      operationId: createBatch
      tags: [batches]
      summary: Create a new commodity batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{batchId}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get batch details
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDetail"
  /batches/{batchId}/lock:
    post:
      operationId: lockBatch
      tags: [batches]
      summary: Lock a batch after all weights confirmed
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Batch locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

  # ============================================================
  # PROCUREMENT
  # ============================================================
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [available, assigned, under_repair, written_off]
        - in: query
          name: assignedTo
          schema:
            type: string
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"

    post:
      operationId: createAsset
      tags: [assets]
      summary: Register a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset details and assignment history
      parameters:
        - in: path
          name: assetId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetail"
  /assets/{assetId}/assign:
    post:
      operationId: assignAsset
      tags: [assets]
      summary: Assign asset to staff member
      parameters:
        - in: path
          name: assetId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAssetBody"
      responses:
        "200":
          description: Asset assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}/return:
    post:
      operationId: returnAsset
      tags: [assets]
      summary: Record asset return
      parameters:
        - in: path
          name: assetId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnAssetBody"
      responses:
        "200":
          description: Return recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"

  # ============================================================
  # ACTIVITY FUNDS
  # ============================================================
...
  /admin/regions:
    get:
      operationId: listRegions
      tags: [admin]
      summary: List administrative regions
      responses:
        "200":
          description: Regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"

    post:
      operationId: createRegion
      tags: [admin]
      summary: Create an administrative region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegionBody"
      responses:
        "201":
          description: Region created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
...
  /admin/roles/{roleId}/permissions:
    patch:
      operationId: updateRolePermissions
      tags: [admin]
      summary: Update permissions for a role
      parameters:
        - in: path
          name: roleId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRolePermissionsBody"
      responses:
        "200":
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
...
components:
  schemas:
...
    CertificationEnrolment:
      type: object
      properties:
        id:
          type: string
        farmerId:
          type: string
        streamId:
          type: string
        streamName:
          type: string
        enrolmentDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, expired, suspended, pending]
      required: [id, farmerId, streamId, streamName, enrolmentDate, status]
    CreateEnrolmentBody:
      type: object
      properties:
        farmerId:
          type: string
        streamId:
          type: string
        enrolmentDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
      required: [farmerId, streamId, enrolmentDate]

    # ============================================================
    # SURVEYS
    # ============================================================
...
    Batch:
      type: object
      properties:
        id:
          type: string
        batchTag:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        cropType:
          type: string
        totalWeightKg:
          type: number
        farmerCount:
          type: integer
        qualifyingStreams:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [open, locked, delivered, received_at_station]
        harvestDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, batchTag, agentId, cropType, totalWeightKg, farmerCount, qualifyingStreams, status, createdAt]
    BatchDetail:
      allOf:
        - $ref: "#/components/schemas/Batch"
        - type: object
          properties:
            farmerContributions:
              type: array
              items:
                type: object
                properties:
                  farmerId:
                    type: string
                  farmerName:
                    type: string
                  grossWeightKg:
                    type: number
                  harvestDate:
                    type: string
                    format: date
                  activeStreams:
                    type: array
                    items:
                      type: string
                required: [farmerId, farmerName, grossWeightKg, harvestDate, activeStreams]
          required: [farmerContributions]
    CreateBatchBody:
      type: object
      properties:
        cropType:
          type: string
        harvestDate:
          type: string
          format: date
      required: [cropType, harvestDate]

    # ============================================================
    # PROCUREMENT
    # ============================================================
...
    Asset:
      type: object
      properties:
        id:
          type: string
        assetCode:
          type: string
        type:
          type: string
          enum: [device, vehicle, scale, equipment, other]
        serialNumber:
          type: string
        make:
          type: string
        model:
          type: string
        purchaseDate:
          type: string
          format: date
        purchaseValue:
          type: number
        currentBookValue:
          type: number
        status:
          type: string
          enum: [available, assigned, under_repair, written_off]
        assignedToUserId:
          type: string
        assignedToName:
          type: string
        photoUrl:
          type: string
        maintenanceDueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, assetCode, type, status, createdAt]
    AssetDetail:
      allOf:
        - $ref: "#/components/schemas/Asset"
        - type: object
          properties:
            assignmentHistory:
              type: array
              items:
                type: object
                properties:
                  userId:
                    type: string
                  userName:
                    type: string
                  assignedAt:
                    type: string
                    format: date-time
                  returnedAt:
                    type: string
                    format: date-time
                  conditionAtHandover:
                    type: string
                  conditionAtReturn:
                    type: string
                required: [userId, userName, assignedAt]
          required: [assignmentHistory]
    CreateAssetBody:
      type: object
      properties:
        type:
          type: string
          enum: [device, vehicle, scale, equipment, other]
        serialNumber:
          type: string
        make:
          type: string
        model:
          type: string
        purchaseDate:
          type: string
          format: date
        purchaseValue:
          type: number
        photoUrl:
          type: string
      required: [type, purchaseDate, purchaseValue]
    AssignAssetBody:
      type: object
      properties:
        userId:
          type: string
        conditionAtHandover:
          type: string
      required: [userId, conditionAtHandover]
    ReturnAssetBody:
      type: object
      properties:
        conditionAtReturn:
          type: string
        photoUrl:
          type: string
        notes:
          type: string
      required: [conditionAtReturn]

    # ============================================================
    # ACTIVITY FUNDS
    # ============================================================
...
    Region:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
        level:
          type: integer
        countryCode:
          type: string
      required: [id, name, level]
    CreateRegionBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: string
        level:
          type: integer
        countryCode:
          type: string
      required: [name, level]
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
      required: [id, name, permissions]
    UpdateRolePermissionsBody:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
      required: [permissions]
...